Dependent query (extended from useQuery options)
Dependent query is useful in a case where there are queries that depend on a preceding query's results. 

Personally, I ran into a problem where I had 5 queries, 4 of which depend on the first query's result. What happens is all 5 queries appear to run synchronously on load and 4 error messages popped up once before running successfully for the second time. The random error messages popping up can be distracting for the end user. This issue can be resolved with the use of dependent queries from React Query's useQuery options. 

https://react-query.tanstack.com/guides/dependent-queries