fix: convert hasura operator with graphql-default (#5040)
* fix: convert hasura operator with graphql-default

When using the Hasura provider with the 'graphql-default' naming convention, snake-case hasura operators (e.g. \_is_null) were not converted to the correct case, causing query errors.
This problem is now fixed. Now the snake_case operator is converted according to the given naming convention (e.g. hasura-default: \_is_null, graphql-default: \_isNull).

close: #4972

* chore: use provider's naming convention

Change to use the namingConvention provided by the provider, as the util is for the provider.
However, the live provider does not currently support namingConvention, so a fixed hasura default is specified.

* chore(hasura): update parameter order for filter utilities

* chore(hasura): fix `generateFilters` usage in list subscriptions

* test(hasura): update `generateFilters` tests

---------

Co-authored-by: Ali Emir Şen <senaliemir@gmail.com>